refactor(useMaterial): useResource拆分
