+++ packages/common/component/MetaTableColumns.vue
@@ -4,11 +4,11 @@
 
 <script setup>
 import { nextTick } from 'vue'
-import { useProperties, useResource, useCanvas } from '@opentiny/tiny-engine-controller'
+import { useProperties, useMaterial, useCanvas } from '@opentiny/tiny-engine-controller'
 import MetaArrayItem from './MetaArrayItem.vue'
 
 const { children: schemaChildren, componentName, props } = useProperties().getSchema()
-const configureMap = useResource().getConfigureMap()
+const configureMap = useMaterial().getConfigureMap()
 const childComponentName =
   configureMap[componentName]?.nestingRule?.childWhitelist?.[0] || schemaChildren?.[0]?.componentName
